fastrtc-groq-voice-agent-main/requirements.txt
aiofiles
aioice
aiortc
annotated-types
anyio
audioop-lts
audioread
av
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
cryptography
decorator
distro
dnspython
fastapi
fastrtc
ffmpy
filelock
flatbuffers
fsspec
google-crc32c
gradio
gradio-client
groq
h11
httpcore
httpx
huggingface-hub
humanfriendly
idna
ifaddr
jinja2
joblib
jsonpatch
jsonpointer
langchain-core
langchain-groq
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lazy-loader
librosa
llvmlite
loguru
markdown-it-py
markupsafe
mdurl
mpmath
msgpack
numba
numpy
onnxruntime
orjson
ormsgpack
packaging
pandas
pillow
platformdirs
pooch
protobuf
pycparser
pydantic
pydantic-core
pydub
python-dotenv
pyee
pygments
pylibsrtp
pyopenssl
pyreadline3
python-dateutil
python-multipart
pytz
pyyaml
requests
requests-toolbelt
rich
ruff
safehttpx
scikit-learn
scipy
semantic-version
shellingham
six
sniffio
soundfile
soxr
standard-aifc
standard-chunk
standard-sunau
starlette
sympy
tenacity
threadpoolctl
tomlkit
tqdm
typer
typing-extensions
typing-inspection
tzdata
urllib3
uvicorn
websockets
win32-setctime
xxhash
zstandard
